requirements.txt
certifi==2020.12.5
chardet==4.0.0
ftfy==5.8
idna==2.10
mailchecker==4.0.3
phonenumbers==8.12.17
python-benedict==0.23.2
python-dateutil==2.8.1
python-fsutil==0.4.0
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
wcwidth==0.2.5
xmltodict==0.12.0
